requirements.txt
Django==4.2.6
django-environ==0.11.2
requests==2.31.0
pillow==10.1.0